Update Key Detection Algorithm (sum of maximum note method)
At each time sample maximum note is chosen. make a zero array and fill the maximum note in the relevant index with respect to the Q transform array. Then sum it up in time's direction. Now merge all the scales together to obtain a single scale. From this sort this 12 element long array. A scale has 12 notes. By removing the lowest occurred 5 values will give you the scale. y considering the notes of that scale we can obtain the key.